FIX improving buffer reference logic

+++ src/animation/JointRegistry.h
@@ -41,7 +41,7 @@ namespace animation {
     protected:
         // Register node instance specific rig
         // @return instance index into joint transform buffer
-        util::BufferReference allocate(uint32_t count);
+        util::BufferReference allocate(size_t count);
         // @return null ref
         util::BufferReference release(util::BufferReference ref);
